add log annotation, check exists when delete

+++ src/main/java/net/vg4/service/CustomerService.java
@@ -3,17 +3,20 @@ package net.vg4.service;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import lombok.extern.slf4j.Slf4j;
 import net.vg4.domain.Customer;
 import net.vg4.domain.User;
 import net.vg4.repository.CustomerRepository;
 
 @Service
 @Transactional
+@Slf4j
 public class CustomerService {
 	@Autowired
 	CustomerRepository customerRepository;
@@ -41,7 +44,15 @@ public class CustomerService {
 	}
 
 	public void delete(Integer id) {
-		customerRepository.delete(id);
+		try {
+			if (customerRepository.exists(id)) {
+				customerRepository.delete(id);
+			} else {
+				log.warn("not exists id", id);
+			}
+		} catch (EmptyResultDataAccessException e) {
+			log.error("Error in delete:");
+		}
 	}
 
 	public Page<Customer> findAll(Pageable pageable) {
